Front end started
Installed React with Next.js. Created a basic homepage that has a brief description of the app, as well as a login form or create event form based on whether you are logged in. Authentication routes added to backend. Front end is connected to backend for login and authenticating the current user.