User auth complete, home page complete
Homepage is responsive to user sessions and automatically switches between the create event form and login form. Made pages for sign up and login that are linked from the nav. Nav switches to profile and logout links when logged in. Made the event management page that you are sent to when you create an event. Made changes to backend routes to require auth to create and update events.