Event page and event management page complete
Added all the details of the event to the event page, a summary of the RSVPs and made the rsvp form. Added an add to calendar component that will has links for Google, Apple, and Outlook calendars. Fixed an issue with the api call to fetch rsvps on the management page. Fixed an issue in the route where it was giving 404 when it should have given a response with all 0's