updated RSVP table in profile, updated rsvp form
RSVP List is now made as a table and has proper styling. Customize the backend route for this table to combine each RSVP with information about its corresponding event. Noticed that RSVP Form and its route were not capture the user information (which meant they didnt show up in profile) so fixed that. Editted RSVP Form to hide when submitted and to allow updating for logged in users. Added ability still connect rsvp to you account if you authenticate after submitting.