Restyled Add to Calendar as a modal
Moved the links to individual calendar apps into a modal pop up window. Each link is part of a card with the name of the app and the icon associated with it (using placeholder logos).  Created a generic modals.css to do styling that could easily be applied somewhere else.