Open in Maps button added. Fixed styling issues
Added a button to the event page that allows you to open the events location in Google Maps. Fixed a styling issue where the page went off the screen on mobile (probably a side effect f the last commit, the footer). Fixed issue where all buttons had an unwanted right margin, and applied the style correctly to the one that needs it ("Copy Event Link" on the event management page).