Added skeletons to items that need to await to load
Added a Skeleton component that will display a bank version of elements while they wait to receive information from the server. Made skeleton types for forms and for page headers. They are modular and support inputs for the item count (how many inner lines or form boxes), for how many, if any fake buttons should be at the bottom, and whether to left justify the content or leave it centered. For some elements like tables, where the length of them varies wildly, I am using a spinner as a placeholder while it loads.

Also added to the App Wrapper to make this work. Making a new context for whether the user is currently loading. Had to make edits to some of the pages to account for the extra state of loading the user versus the user simply being null of filled.

Made adjustments on every page, that isn't static, in order to render a page skeleton for any dynamic elements.