Handled errantly navigating to profile while logged out
Through normal navigation you cannot get to the profile screen while logged out. This change accounts for users navigating through the back button in their browser to get to the profile page while logged out. This was throwing an error before, but now it will render a basic page telling you to login or sign up to view the page.