Updated icons and colors. Restyled RSVP stats
Updated calendar icons and github icons from their placeholders. Made and implemented a new logo for the site. Added edit, view and copy icons. Replaced text links of "manage" and "view" in profile page. Added the cop and view icons to the main buttons on the event management page. Changed the colors for the site. Propagated color changes throughout the css and also updated most literal colors to refer to a variable. Restyled the RSVP stats to have a capsule shape around it.