Fixed bug with scroll bar
Fixed a bug where the horizontal scroll bar was not flush with the edge of the page. This was caused when I put styling to add padding to edge of the site on small screens, I put it on the body instead of doing it on an inner element. So I moved that styling to the appropriate class.